download/requirements_downloader.txt
huggingface_hub>=0.16.0
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.31.0 